+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.8.8-rc.1
+
+- Eigener Nextcloud-Adminabschnitt für die ausschließlich appbezogene Raumverwaltung.
+- Öffentliche Projekt-, Quellcode- und Fehlerkanäle ergänzt.
+
 ## 0.8.7-rc.1
 
 - Erster reproduzierbarer Staging-Releasekandidat für Nextcloud 34 und PHP ab 8.3.